Also rename "library" file
